+++ src/objects.js
@@ -86,7 +86,7 @@ AlignmentMorph, Process, WorldMap, copyCanvas, useBlurredShadows*/
 
 /*jshint esversion: 6*/
 
-modules.objects = '2021-July-09';
+modules.objects = '2021-July-12';
 
 var SpriteMorph;
 var StageMorph;
@@ -9522,6 +9522,10 @@ StageMorph.prototype.globalVariables
 
 StageMorph.prototype.inheritedVariableNames = function () {
     return [];
+};
+
+StageMorph.prototype.deletableVariableNames = function () {
+    return this.variables.allNames();
 };
 
 StageMorph.prototype.allLocalVariableNames